requirements.txt
click>=8.0
anthropic>=0.20
python-dotenv>=1.0
